Update and rename d.py to jarvis1.0.py
